gcf/python/telemetry-to-bigquery/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==2.31.0
